pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cm-data-transformation"
version = "0.1.0"
description = "A Python-based data transformation tool with spatial SQL operations."
authors = [
    { name="CleverMaps", email="info@clevermaps.com" }
]
readme = "README.md"
requires-python = ">=3.7"
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
dependencies = [
    "duckdb",
    "jinja2",
    "pyyaml",
    "sqlalchemy",
    "duckdb-engine"
]

[tool.setuptools.packages.find]
where = ["src"]
